Rename PapiClient behavior test namespace
