+++ src/polaris-api-csharpTests/Methods/IntegrationTestBase.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Linq;
 using Clc.Polaris.Api;
-using PapiClientType = Clc.Polaris.Api.PapiClient;
 using Clc.Polaris.Api.Configuration;
 using Microsoft.Extensions.Configuration;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -39,7 +38,7 @@ namespace Clc.Polaris.Api.Tests
                 Assert.Inconclusive(IntegrationTestRequirements.MissingIntegrationConfigurationMessage);
             }
 
-            Papi = new PapiClientType(papiSettings!);
+            Papi = new PapiClient(papiSettings!);
             PapiSettings = papiSettings!;
             Settings = testSettings!;
         }
