+++ src/polaris-api-csharpTests/PapiClientBehavior/PreformatRestRequestIdempotencyTests.cs
@@ -6,13 +6,12 @@ using Clc.Polaris.Api.Tests;
 using Clc.Polaris.Api.Tests.TestInfrastructure;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Clc.Polaris.Api.Tests.PapiClient
+namespace Clc.Polaris.Api.Tests.PapiClientBehavior
 {
     [TestClass]
     [UnitCategory]
     public class PreformatRestRequestIdempotencyTests : PapiClientTestBase
     {
-
         [TestMethod]
         public void PreformatRestRequest_WhenCalledTwice_ReplacesPapiHeadersWithoutDuplicatingOrChangingBody()
         {
